tests/test_predictive_modeling.py
import unittest
from unittest.mock import patch
from src.predictive_modeling import evaluate_performance, evaluate_safety, evaluate_efficiency

class TestPredictiveModeling(unittest.TestCase):
    @patch('src.predictive_modeling.load_model')
    def test_evaluate_performance(self, mock_load_model):
        mock_model = mock_load_model.return_value
        mock_model.predict.return_value = [85]
        design = {
            "parameters": [1, 2, 3, 4, 5]
        }
        performance_score = evaluate_performance(design)
        self.assertEqual(performance_score, 85)

    @patch('src.predictive_modeling.load_model')
    def test_evaluate_safety(self, mock_load_model):
        mock_model = mock_load_model.return_value
        mock_model.predict.return_value = [90]
        design = {
            "parameters": [1, 2, 3, 4, 5]
        }
        safety_score = evaluate_safety(design)
        self.assertEqual(safety_score, 90)

    @patch('src.predictive_modeling.load_model')
    def test_evaluate_efficiency(self, mock_load_model):
        mock_model = mock_load_model.return_value
        mock_model.predict.return_value = [75]
        design = {
            "parameters": [1, 2, 3, 4, 5]
        }
        efficiency_score = evaluate_efficiency(design)
        self.assertEqual(efficiency_score, 75)

if __name__ == '__main__':
    unittest.main()
